# Use out-of-proc MSBuild evaluation