--- a/src/Microsoft.Tye.Extensions/Microsoft.Tye.Extensions.csproj
+++ b/src/Microsoft.Tye.Extensions/Microsoft.Tye.Extensions.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>