--- a/src/tye-diag-agent/tye-diag-agent.csproj
+++ b/src/tye-diag-agent/tye-diag-agent.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Worker">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>Microsoft.Tye</RootNamespace>
   </PropertyGroup>