… test/E2ETest/Microsoft.Tye.E2ETests.csproj
… test/Test.Infrastructure/Test.Infrastructure.csproj
… test/UnitTests/Microsoft.Tye.UnitTests.csproj