Fixes https://github.com/dotnet/tye/issues/594

I have tested to ensure that the original issue where installing a 5.0 SDK would break apps has been fixed with this change. With this change, supporting new TFMs wouldn't require tye itself to be re-targeted.

So far testing the perf with frontend-backend sample gives:

Before:
backend.csproj
Evaluation Took: 1500.9242ms
frontend.csproj
Evaluation Took: 463.2118ms

After:
backend.csproj
Evaluation Took: 1733.3304ms
frontend.csproj
Evaluation Took: 1766.8971ms

So there is a performance penalty with running out of proc, especially with additional projects.