@davidfowl 2020-09-22
I'm not sure I want to pay this price. I'd like to discuss different approaches, like possibly doing more computation of all projects in a single pass rather than trigger an invocation per project.