@JunTaoLuo 2020-09-25
Updated numbers:

After:
backend.csproj and frontend.csproj: 2207.3488ms

I'll collect some numbers with more projects tomorrow.